use floats for prices

+++ src/networkMarket/interfaces/Store.java
@@ -16,11 +16,11 @@ public interface Store extends Remote {
 
     List<Item> getItems(User user) throws RemoteException, UserException;
 
-    Item addItem(String name, double price, User user) throws RemoteException, UserException;
+    Item addItem(String name, float price, User user) throws RemoteException, UserException;
 
     void buyItem(Item item, User user) throws RemoteException, RejectedException, UserException;
 
-    void wishItem(String name, double price, User user) throws RemoteException, UserException;
+    void wishItem(String name, float price, User user) throws RemoteException, UserException;
 
     List<Wish> getMyWishes(User user) throws RemoteException, UserException;
